Initial support for aarch64 based cluter
This commits introduces a lot of breaking changes relative
to the ClearLinux based upstream in order to support aarch64 manifests.

Also, moving to Ubuntu as the basis for node-setup. The Clearlinux
Kubernetes reference stack had a sane and well maintained framework for
managing a typical stack, and is a great starting point.  In the future,
plan to node-selects for manifests which are architecture specific, and
may look at augmenting existing setup scripts in the upstream to support
both architectures/hosts.

Signed-off-by: Eric Ernst <eric@amperecomputing.com>