cni: create multi-arch image for flannel
v0.12.0 flannel does not have multi-arch support, but the rest of the
images needed for running canal are already multiarch.

Created a multi-arch image for flannel @ egernst/coreos-flannel:v0.11.0
following directions @
https://www.docker.com/blog/multi-arch-build-and-images-the-simple-way/
(just using the existing arch specific images).

With this in place, we can start a bare-bones k8s cluster with
networking on a mixed cluster (both aarch64 and x86_64).

Signed-off-by: Eric Ernst <eric@amperecomputing.com>